Merging commits that were staged before query code was merged into dev

+++ src/queries-Robert/queryRunners/delete.js
@@ -30,21 +30,21 @@ let execDelete = async (query, dataBase, commitResults) => {
         queryInfo.collection = qp.getCollection(query, 'delete');
         queryInfo.wheres = qp.getWheres(query);
         queryInfo.selectFields = ['*'];
-        const payload = selectDb.getDataForSelect(queryInfo, dataBase); // use getDataForSelect to determine what we need to delete
+        let payload = selectDb.getDataForSelect(queryInfo, dataBase); // use getDataForSelect to determine what we need to delete
         await payload.then((data) => {
             if (data && commitResults){
                 if (!queryInfo.wheres && queryInfo.collection.indexOf('/') > 0) {
                     // unfiltered: delete from users.userId
-                    deleteDb.deleteObject(queryInfo.collection, dataBase)
-                    // console.log('Deleting... ', queryInfo.collection)
+                    deleteDb.deleteObject(queryInfo.collection, dataBase);
                 } else {
                     // Use select payload to determine deletes:
                     // entire col: delete from users;
                     // OR filtered: delete from users where age > x;
                     Object.keys(data).forEach(objKey => {
-                        const path = queryInfo.collection + "/" + objKey;
-                        // console.log('Deleting... ', path);
-                        deleteDb.deleteObject(path, dataBase);
+                        if (data[objKey]) {
+                            const path = queryInfo.collection + "/" + objKey;
+                            deleteDb.deleteObject(path, dataBase);
+                        }
                     })
                 }
             }
